move test back to core suite
